refactor: extract header-rendering logic into TableHeaderRenderer
